Servicicio de facturación terminado

+++ TuyaPagos/TuyaPagos.Domain/Entities/DetalleFactura.cs
@@ -2,11 +2,11 @@
 {
     public class DetalleFactura
     {
-        public long Id { get; set; }
-        public long FacturaId { get; set; }
+        public int Id { get; set; }
+        public int FacturaId { get; set; }
         public virtual Factura Factura { get; set; }
         public int Cantidad { get; set; }
-        public long ProductoId { get; set; }
+        public int ProductoId { get; set; }
         public virtual Producto Producto { get; set; }
         public long ValorBruto { get; set; }
         public int Impuesto { get; set; }
