+++ TuyaPagos/TuyaPagos.Domain/Entities/Pedido.cs
@@ -2,10 +2,10 @@
 {
     public class Pedido
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime Fecha { get; set; }
         public string Numero { get; set; }
-        public long FacturaId { get; set; }
+        public int FacturaId { get; set; }
         public virtual Factura Factura { get; set; }
         public string Direccion { get; set; }
         public string Ciudad { get; set; }
